Fix warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@27392 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/osm/gpb.h
@@ -140,7 +140,7 @@ static unsigned int ReadVarUInt32(GByte** ppabyData)
 #define READ_SIZE(pabyData, pabyDataLimit, nSize) \
     { \
         READ_VARUINT32(pabyData, pabyDataLimit, nSize); \
-        if (CHECK_OOB && nSize > pabyDataLimit - pabyData) GOTO_END_ERROR; \
+        if (CHECK_OOB && nSize > (unsigned int)(pabyDataLimit - pabyData)) GOTO_END_ERROR; \
     }
 
 /************************************************************************/
